+++ src/azure-cli/azure/cli/command_modules/resource/custom.py
@@ -39,6 +39,9 @@ from knack.util import CLIError
 
 from ._validators import MSI_LOCAL_ID
 
+from msrest.serialization import Serializer
+from msrest.pipeline import SansIOHTTPPolicy
+
 logger = get_logger(__name__)
 
 
@@ -296,7 +299,7 @@ def _remove_comments_from_json(template):
 
 
 # pylint: disable=too-many-locals, too-many-statements, too-few-public-methods
-def _deploy_arm_template_unmodified(cli_ctx, resource_group_name, template_file=None,
+def _deploy_arm_template_core_unmodified(cli_ctx, resource_group_name, template_file=None,
                                     template_uri=None, deployment_name=None, parameters=None,
                                     mode=None, rollback_on_error=None, validate_only=False, no_wait=False):
     DeploymentProperties, TemplateLink, OnErrorDeployment = get_sdk(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES,
@@ -331,49 +334,12 @@ def _deploy_arm_template_unmodified(cli_ctx, resource_group_name, template_file=
 
     smc = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES)
 
-    class JsonCTemplate(object):
-        def __init__(self, template_as_bytes):
-            self.template_as_bytes = template_as_bytes
-
-    from msrest.serialization import Serializer
-
-    class MySerializer(Serializer):
-        def body(self, data, data_type, **kwargs):
-            if data_type == 'Deployment':
-                # Be sure to pass a DeploymentProperties
-                template = data.properties.template
-                if template:
-                    data.properties.template = None
-                    data_as_dict = data.serialize()
-                    data_as_dict["properties"]["template"] = JsonCTemplate(template)
-                    return data_as_dict
-            return super(MySerializer, self).body(data, data_type, **kwargs)
     deployments_operation_group = smc.deployments  # This solves the multi-api for you
 
     # pylint: disable=protected-access
-    deployments_operation_group._serialize = MySerializer(
+    deployments_operation_group._serialize = JSONSerializer(
         deployments_operation_group._serialize.dependencies
     )
-
-    # Now, you have a serializer that keeps a weird class inside it, you need to explain to the HTTP pipeline how to translate that into bytes
-    from msrest.pipeline import SansIOHTTPPolicy
-
-    class JsonCTemplatePolicy(SansIOHTTPPolicy):
-        def on_request(self, request, **kwargs):
-            http_request = request.http_request
-            logger.info(http_request.data)
-            if (getattr(http_request, 'data', {}) or {}).get('properties', {}).get('template'):
-                template = http_request.data["properties"]["template"]
-                if not isinstance(template, JsonCTemplate):
-                    raise ValueError()
-
-                del http_request.data["properties"]["template"]
-                # templateLink nad template cannot exist at the same time in deployment_dry_run mode
-                if "templateLink" in http_request.data["properties"].keys():
-                    del http_request.data["properties"]["templateLink"]
-                partial_request = json.dumps(http_request.data)
-
-                http_request.data = partial_request[:-2] + ", template:" + template.template_as_bytes + r"}}"
 
     # Plug this as default HTTP pipeline
     from msrest.pipeline import Pipeline
@@ -398,6 +364,115 @@ def _deploy_arm_template_unmodified(cli_ctx, resource_group_name, template_file=
     if validate_only:
         return sdk_no_wait(no_wait, deployments_operation_group.validate, resource_group_name, deployment_name, properties)
     return sdk_no_wait(no_wait, deployments_operation_group.create_or_update, resource_group_name, deployment_name, properties)
+
+
+class JsonCTemplate(object):
+    def __init__(self, template_as_bytes):
+        self.template_as_bytes = template_as_bytes
+
+
+class JSONSerializer(Serializer):
+    def body(self, data, data_type, **kwargs):
+        if data_type == 'Deployment':
+            # Be sure to pass a DeploymentProperties
+            template = data.properties.template
+            if template:
+                data.properties.template = None
+                data_as_dict = data.serialize()
+                data_as_dict["properties"]["template"] = JsonCTemplate(template)
+                return data_as_dict
+        return super(JSONSerializer, self).body(data, data_type, **kwargs)
+
+
+class JsonCTemplatePolicy(SansIOHTTPPolicy):
+    def on_request(self, request, **kwargs):
+        http_request = request.http_request
+        logger.info(http_request.data)
+        if (getattr(http_request, 'data', {}) or {}).get('properties', {}).get('template'):
+            template = http_request.data["properties"]["template"]
+            if not isinstance(template, JsonCTemplate):
+                raise ValueError()
+
+            del http_request.data["properties"]["template"]
+            # templateLink nad template cannot exist at the same time in deployment_dry_run mode
+            if "templateLink" in http_request.data["properties"].keys():
+                del http_request.data["properties"]["templateLink"]
+            partial_request = json.dumps(http_request.data)
+
+            http_request.data = partial_request[:-2] + ", template:" + template.template_as_bytes + r"}}"
+
+
+def _deploy_arm_template_unmodified(cli_ctx,
+                                    scope_type=None, resource_group_name=None, management_group_id=None,
+                                    template_file=None, template_uri=None,
+                                    deployment_name=None, deployment_location=None,
+                                    parameters=None, mode=None, validate_only=False,
+                                    no_wait=False):
+    DeploymentProperties, TemplateLink = get_sdk(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES,
+                                                 'DeploymentProperties', 'TemplateLink', mod='models')
+    template = None
+    template_link = None
+    template_obj = None
+    if template_uri:
+        template_link = TemplateLink(uri=template_uri)
+        template_content = _urlretrieve(template_uri).decode('utf-8')
+        template_obj = _remove_comments_from_json(template_content)
+    else:
+        template_content = read_file_content(template_file)
+        template_obj = _remove_comments_from_json(template_content)
+
+    template_param_defs = template_obj.get('parameters', {})
+    template_obj['resources'] = template_obj.get('resources', [])
+    parameters = _process_parameters(template_param_defs, parameters) or {}
+    parameters = _get_missing_parameters(parameters, template_obj, _prompt_for_parameters)
+
+    template = json.loads(json.dumps(template))
+    parameters = json.loads(json.dumps(parameters))
+
+    properties = DeploymentProperties(template=template, template_link=template_link,
+                                      parameters=parameters, mode=mode)
+
+    smc = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES)
+
+    deployments_operation_group = smc.deployments  # This solves the multi-api for you
+
+    # pylint: disable=protected-access
+    deployments_operation_group._serialize = JSONSerializer(
+        deployments_operation_group._serialize.dependencies
+    )
+
+    # Plug this as default HTTP pipeline
+    from msrest.pipeline import Pipeline
+    from msrest.pipeline.requests import (
+        RequestsCredentialsPolicy,
+        RequestsPatchSession,
+        PipelineRequestsHTTPSender
+    )
+    from msrest.universal_http.requests import RequestsHTTPSender
+
+    smc.config.pipeline = Pipeline(
+        policies=[
+            JsonCTemplatePolicy(),
+            smc.config.user_agent_policy,
+            RequestsPatchSession(),
+            smc.config.http_logger_policy,
+            RequestsCredentialsPolicy(smc.config.credentials)
+        ],
+        sender=PipelineRequestsHTTPSender(RequestsHTTPSender(smc.config))
+    )
+
+    smc = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES)
+
+    if scope_type == 'Subscription':
+        return _deploy_at_subscription_scope(smc, deployment_name, deployment_location, properties, validate_only, no_wait)
+
+    if scope_type == 'ResourceGroup':
+        return _deploy_at_resource_group(smc, resource_group_name, deployment_name, properties, validate_only, no_wait)
+
+    if scope_type == 'ManagementGroup':
+        return _deploy_at_management_group(smc, management_group_id, deployment_name, deployment_location, properties, validate_only, no_wait)
+
+    return _deploy_at_tenant_scope(smc, deployment_name, deployment_location, properties, validate_only, no_wait)
 
 
 def _deploy_arm_template(cli_ctx,
@@ -429,55 +504,49 @@ def _deploy_arm_template(cli_ctx,
     properties = DeploymentProperties(template=template, template_link=template_link,
                                       parameters=parameters, mode=mode)
 
-    if scope_type == 'Subscription':
-        return _deploy_at_subscription_scope(cli_ctx, deployment_name, deployment_location, properties, validate_only, no_wait)
-
-    if scope_type == 'ResourceGroup':
-        return _deploy_at_resource_group(cli_ctx, resource_group_name, deployment_name, properties, validate_only, no_wait)
-
-    if scope_type == 'ManagementGroup':
-        return _deploy_at_management_group(cli_ctx, management_group_id, deployment_name, deployment_location, properties, validate_only, no_wait)
-
-    return _deploy_at_tenant_scope(cli_ctx, deployment_name, deployment_location, properties, validate_only, no_wait)
-
-
-def _deploy_at_subscription_scope(cli_ctx, deployment_name=None, deployment_location=None, deployment_properties=None, validate_only=False, no_wait=False):
     smc = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES)
 
+    if scope_type == 'Subscription':
+        return _deploy_at_subscription_scope(smc, deployment_name, deployment_location, properties, validate_only, no_wait)
+
+    if scope_type == 'ResourceGroup':
+        return _deploy_at_resource_group(smc, resource_group_name, deployment_name, properties, validate_only, no_wait)
+
+    if scope_type == 'ManagementGroup':
+        return _deploy_at_management_group(smc, management_group_id, deployment_name, deployment_location, properties, validate_only, no_wait)
+
+    return _deploy_at_tenant_scope(smc, deployment_name, deployment_location, properties, validate_only, no_wait)
+
+
+def _deploy_at_subscription_scope(mgmt_client, deployment_name=None, deployment_location=None, deployment_properties=None, validate_only=False, no_wait=False):
     if validate_only:
-        return sdk_no_wait(no_wait, smc.deployments.validate_at_subscription_scope,
+        return sdk_no_wait(no_wait, mgmt_client.deployments.validate_at_subscription_scope,
                            deployment_name, deployment_properties, deployment_location)
-    return sdk_no_wait(no_wait, smc.deployments.create_or_update_at_subscription_scope,
+    return sdk_no_wait(no_wait, mgmt_client.deployments.create_or_update_at_subscription_scope,
                        deployment_name, deployment_properties, deployment_location)
 
 
-def _deploy_at_resource_group(cli_ctx, resource_group_name=None, deployment_name=None, deployment_properties=None, validate_only=False, no_wait=False):
-    smc = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES)
-
+def _deploy_at_resource_group(mgmt_client, resource_group_name=None, deployment_name=None, deployment_properties=None, validate_only=False, no_wait=False):
     if validate_only:
-        return sdk_no_wait(no_wait, smc.deployments.validate, resource_group_name,
+        return sdk_no_wait(no_wait, mgmt_client.deployments.validate, resource_group_name,
                            deployment_name, deployment_properties)
-    return sdk_no_wait(no_wait, smc.deployments.create_or_update, resource_group_name,
+    return sdk_no_wait(no_wait, mgmt_client.deployments.create_or_update, resource_group_name,
                        deployment_name, deployment_properties)
 
 
-def _deploy_at_management_group(cli_ctx, management_group_id=None, deployment_name=None, deployment_location=None, deployment_properties=None, validate_only=False, no_wait=False):
-    smc = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES)
-
+def _deploy_at_management_group(mgmt_client, management_group_id=None, deployment_name=None, deployment_location=None, deployment_properties=None, validate_only=False, no_wait=False):
     if validate_only:
-        return sdk_no_wait(no_wait, smc.deployments.validate_at_management_group_scope,
+        return sdk_no_wait(no_wait, mgmt_client.deployments.validate_at_management_group_scope,
                            management_group_id, deployment_name, deployment_properties, deployment_location)
-    return sdk_no_wait(no_wait, smc.deployments.create_or_update_at_management_group_scope,
+    return sdk_no_wait(no_wait, mgmt_client.deployments.create_or_update_at_management_group_scope,
                        management_group_id, deployment_name, deployment_properties, deployment_location)
 
 
-def _deploy_at_tenant_scope(cli_ctx, deployment_name=None, deployment_location=None, deployment_properties=None, validate_only=False, no_wait=False):
-    smc = get_mgmt_service_client(cli_ctx, ResourceType.MGMT_RESOURCE_RESOURCES)
-
+def _deploy_at_tenant_scope(mgmt_client, deployment_name=None, deployment_location=None, deployment_properties=None, validate_only=False, no_wait=False):
     if validate_only:
-        return sdk_no_wait(no_wait, smc.deployments.validate_at_tenant_scope,
+        return sdk_no_wait(no_wait, mgmt_client.deployments.validate_at_tenant_scope,
                            deployment_name, deployment_properties, deployment_location)
-    return sdk_no_wait(no_wait, smc.deployments.create_or_update_at_tenant_scope,
+    return sdk_no_wait(no_wait, mgmt_client.deployments.create_or_update_at_tenant_scope,
                        deployment_name, deployment_properties, deployment_location)
 
 
@@ -988,7 +1057,7 @@ def deploy_arm_template(cmd, resource_group_name,
                         template_file=None, template_uri=None, deployment_name=None,
                         parameters=None, mode=None, rollback_on_error=None, no_wait=False, handle_extended_json_format=False):
     if handle_extended_json_format:
-        return _deploy_arm_template_unmodified(cmd.cli_ctx, resource_group_name, template_file, template_uri,
+        return _deploy_arm_template_core_unmodified(cmd.cli_ctx, resource_group_name, template_file, template_uri,
                                                deployment_name, parameters, mode, rollback_on_error, no_wait=no_wait)
 
     return _deploy_arm_template_core(cmd.cli_ctx, resource_group_name, template_file, template_uri,
@@ -999,7 +1068,15 @@ def deploy_arm_template_at_scope(cmd,
                                  scope_type=None, resource_group_name=None, management_group_id=None,
                                  template_file=None, template_uri=None,
                                  deployment_name=None, deployment_location=None,
-                                 parameters=None, no_wait=False):
+                                 parameters=None, no_wait=False, handle_extended_json_format=None):
+
+    if handle_extended_json_format:
+        _deploy_arm_template_unmodified(cmd.cli_ctx,
+                                        scope_type, resource_group_name, management_group_id,
+                                        template_file, template_uri,
+                                        deployment_name, deployment_location,
+                                        parameters, 'Incremental', no_wait=no_wait)
+
     return _deploy_arm_template(cmd.cli_ctx,
                                 scope_type, resource_group_name, management_group_id,
                                 template_file, template_uri,